+++ apps/web/components/layout/header-icon/RegionLabel.tsx
@@ -3,11 +3,13 @@
 import { Button } from '@repo/ui/components';
 import { useQuery } from '@tanstack/react-query';
 
-import { fetchUserRegion } from '@/lib/query/client';
+import { fetchUserRegion } from '@/services/user/client';
+
+import { USER_REGION_QUERY_KEY } from '@/constants';
 
 const RegionLabel = () => {
   const { data: region } = useQuery<string | null>({
-    queryKey: ['user-region'],
+    queryKey: USER_REGION_QUERY_KEY,
     queryFn: fetchUserRegion,
   });
