+++ apps/web/constants/index.ts
@@ -3,3 +3,4 @@ export * from './routes';
 export * from './products';
 export * from './storage-bucket';
 export * from './navigation';
+export * from './query-key';
