+++ apps/web/constants/query-key.ts
@@ -1,0 +1,3 @@
+// 공통 queryKey 상수
+
+export const USER_REGION_QUERY_KEY = ['user-region'] as const;
