+++ apps/web/constants/routes/api.ts
@@ -4,4 +4,5 @@ export const API_ROUTES = {
   PRODUCTS: `${API_BASE}/products`,
   IMAGES: `${API_BASE}/images`,
   BIDS: `${API_BASE}/bids`,
+  USER: `${API_BASE}/user`,
 } as const;
