+++ apps/web/lib/query/server/index.ts
@@ -1,0 +1,2 @@
+export { createServerQueryClient } from './instance';
+export * from './userQueryServer';
