+++ apps/web/services/user/client/index.ts
@@ -1,0 +1,1 @@
+export * from './fetchUserRegion';
